Move string testing code from MatasanoUtil into new StringTesting file
